fix(provision): use pvc uuid as the PV name
In 0.7.0, the PV Name was set to use the PVC namespace,
name and a unique hash. This lead to some breackages
with usecases where PVCs were auto-generated like in
StatefulSets where the length of the PV Names can
be more than 64 characters.

It is also possible that PVC Name will contain some special
characters that are not honored by iSCSI IQN name, which is
derived from the PV Name.

Reverting back to old name generation scheme:
 pvc-<pvc-uuid>

Signed-off-by: kmova <kiran.mova@openebs.io>