feat(jiva): env to enable/disable replica patch
In some cases, users would like to disable setting
the patch on jiva replica. These are clusters which
guarantee that only one node is out of the cluster
at any given time and when it comes back, it can
come back with an new hostname, making the replica
deployment unschedulable.

The setting can be controlled via ENV variable.

Note: The patch application was being done in the
volume read operation and hence can't be passed
via storage class at the moment.

Signed-off-by: kmova <kiran.mova@mayadata.io>